requirements.txt
python-binance>=1.0.19
pandas>=2.1.4
numpy>=1.26.2
python-dotenv>=1.0.0
schedule>=1.2.1
PyQt5>=5.15.10
matplotlib>=3.8.0
PyQtChart>=5.15.6
cryptography>=41.0.0 